backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
]
